requirements.txt
# Core data processing
pyspark==3.5.3
pandas>=2.2.0
numpy>=2.0.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.2
plotly>=5.24.0

# Machine Learning
scikit-learn>=1.5.0

# Streaming and Delta Lake
kafka-python3>=3.0.0
delta-spark==3.2.1
pyarrow>=17.0.0

# Jupyter environment
jupyter>=1.1.1
notebook>=7.2.0
ipykernel>=6.29.0
